Refactored ClientActor and ClientObject

+++ jgcf/src/main/java/com/jenjinstudios/world/ClientObject.java
@@ -12,7 +12,7 @@ public class ClientObject
 	/** The vector2D in the world at which the object is located. */
 	private Vector2D vector2D;
 	/** The direction in which this object is facing. */
-	private float direction;
+	private double direction;
 	/** The ID number of this object. */
 	private int id = Integer.MIN_VALUE;
 
@@ -32,7 +32,7 @@ public class ClientObject
 	 *
 	 * @return The direction in which this object is facing.
 	 */
-	public float getDirection()
+	public double getDirection()
 	{
 		return direction;
 	}
@@ -42,7 +42,7 @@ public class ClientObject
 	 *
 	 * @param direction The new direction for this object to face.
 	 */
-	public void setDirection(float direction)
+	public void setDirection(double direction)
 	{
 		this.direction = direction;
 	}
